+++ contrib/opencensus-ext-stackdriver/setup.py
@@ -44,7 +44,7 @@ setup(
     ],
     extras_require={},
     license='Apache-2.0',
-    packages=find_packages(exclude=('tests',)),
+    packages=find_packages(exclude=('examples', 'tests',)),
     namespace_packages=[],
     url='https://github.com/census-instrumentation/opencensus-python/tree/master/contrib/opencensus-ext-stackdriver',
     zip_safe=False,
